Implement soil albedo function with some incorrect functions
PAR and NIR albedo moved to aux vars, and replace by PAR and NIR dry/wet
albedos. Currently, PAR and NIR are used to hold the moisture volume fractions
of the top of the soil. This causes repeat calculation, so a var for it
should be added to aux vars. The function that calculates the liquid volume
fraction currently does so from the first level to the interpolated surface.
This should be fixed.